Updated l2fprod to better embrace the Java way of doing things. Use of ServiceLoader for finding Editors and Renderers.

+++ src/com/l2fprod/common/swing/renderer/ColorCellRenderer.java
@@ -53,7 +53,7 @@ public class ColorCellRenderer extends DefaultCellRenderer {
     @Override
     protected String convertToString(Object value) {
         if (value instanceof Integer) {
-            value = new Color(((Integer) value).intValue());
+            value = new Color(((Integer) value));
         }
         if (!(value instanceof Color)) {
             return null;
